Added Node interface and Category class
A new Node interface is added to allow category and assignment class to share the getScore method, and to allow both to be kept in the same list.

The Category class is added, which is built to hold assignments and to have a specific weight on it. This weight will be applied to the score of the category.